requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
cloudinary==1.39.0
python-multipart==0.0.9
PyJWT==2.8.0
pytz==2024.1
openpyxl==3.1.2
itsdangerous>=2.1.0
pandas
openpyxl

